Separate cookie handling from banner to separate class and support multiple required groups
